Extract the relation card and chip shells from the field widget triplets
FieldWidget{Relation,MorphRelation,JunctionRelation}Card repeated the same
provider stack, read-only wiring, show-more pagination and list rendering
(~150 lines each) and only differed in how the related records and their
object name were derived; the three *Field components repeated the same
chip container around the same derivations.

FieldWidgetRelationRecordsCard and FieldWidgetRelationRecordChips now own
the shared rendering and take the derived records. The six existing
components become adapters that only derive records, and the junction
derivation they shared moves into useFieldWidgetJunctionRelationRecords.
The junction card stays forced read-only through the shell's isReadOnly
prop, for the same reason as before: the list item's detach/delete logic
assumes a direct relation.
