+++ packages/twenty-front/src/modules/page-layout/widgets/field/components/FieldWidgetMorphRelationField.tsx
@@ -1,22 +1,7 @@
-import { RecordChip } from '@/object-record/components/RecordChip';
 import { useGetMorphRelationRelatedRecordsWithObjectNameSingular } from '@/object-record/record-field-list/record-detail-section/relation/components/hooks/useGetMorphRelationRelatedRecordsWithObjectNameSingular';
 import { type FieldDefinition } from '@/object-record/record-field/ui/types/FieldDefinition';
 import { type FieldMorphRelationMetadata } from '@/object-record/record-field/ui/types/FieldMetadata';
-import { SidePanelProvider } from '@/ui/layout/side-panel/contexts/SidePanelContext';
-import { styled } from '@linaria/react';
-import { themeCssVariables } from 'twenty-ui/theme-constants';
-
-const StyledContainer = styled.div`
-  box-sizing: border-box;
-  padding: ${themeCssVariables.spacing[1]};
-  width: 100%;
-`;
-
-const StyledRelationChipsContainer = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  gap: ${themeCssVariables.spacing[1]};
-`;
+import { FieldWidgetRelationRecordChips } from '@/page-layout/widgets/field/components/FieldWidgetRelationRecordChips';
 
 type FieldWidgetMorphRelationFieldProps = {
   fieldDefinition: FieldDefinition<FieldMorphRelationMetadata>;
@@ -29,12 +14,10 @@ export const FieldWidgetMorphRelationField = ({
   recordId,
   isInSidePanel,
 }: FieldWidgetMorphRelationFieldProps) => {
-  const fieldMetadata = fieldDefinition.metadata;
-
   const recordsWithObjectNameSingular =
     useGetMorphRelationRelatedRecordsWithObjectNameSingular({
       recordId,
-      morphRelations: fieldMetadata.morphRelations,
+      morphRelations: fieldDefinition.metadata.morphRelations,
     });
 
   if (recordsWithObjectNameSingular.length === 0) {
@@ -42,18 +25,12 @@ export const FieldWidgetMorphRelationField = ({
   }
 
   return (
-    <SidePanelProvider value={{ isInSidePanel }}>
-      <StyledContainer>
-        <StyledRelationChipsContainer>
-          {recordsWithObjectNameSingular.map((morphItem, index) => (
-            <RecordChip
-              key={morphItem.value?.id ?? index}
-              objectNameSingular={morphItem.objectNameSingular}
-              record={morphItem.value}
-            />
-          ))}
-        </StyledRelationChipsContainer>
-      </StyledContainer>
-    </SidePanelProvider>
+    <FieldWidgetRelationRecordChips
+      relationRecords={recordsWithObjectNameSingular.map((morphItem) => ({
+        record: morphItem.value,
+        objectNameSingular: morphItem.objectNameSingular,
+      }))}
+      isInSidePanel={isInSidePanel}
+    />
   );
 };
