+++ packages/twenty-front/src/modules/page-layout/widgets/field/components/FieldWidgetRelationField.tsx
@@ -1,22 +1,8 @@
-import { RecordChip } from '@/object-record/components/RecordChip';
 import { type FieldDefinition } from '@/object-record/record-field/ui/types/FieldDefinition';
 import { type FieldRelationMetadata } from '@/object-record/record-field/ui/types/FieldMetadata';
-import { SidePanelProvider } from '@/ui/layout/side-panel/contexts/SidePanelContext';
-import { styled } from '@linaria/react';
+import { type ObjectRecord } from '@/object-record/types/ObjectRecord';
+import { FieldWidgetRelationRecordChips } from '@/page-layout/widgets/field/components/FieldWidgetRelationRecordChips';
 import { isDefined } from 'twenty-shared/utils';
-import { themeCssVariables } from 'twenty-ui/theme-constants';
-
-const StyledContainer = styled.div`
-  box-sizing: border-box;
-  padding: ${themeCssVariables.spacing[1]};
-  width: 100%;
-`;
-
-const StyledRelationChipsContainer = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  gap: ${themeCssVariables.spacing[1]};
-`;
 
 type FieldWidgetRelationFieldProps = {
   fieldDefinition: FieldDefinition<FieldRelationMetadata>;
@@ -42,34 +28,17 @@ export const FieldWidgetRelationField = ({
     return null;
   }
 
-  if (isOneToMany) {
-    return (
-      <SidePanelProvider value={{ isInSidePanel }}>
-        <StyledContainer>
-          <StyledRelationChipsContainer>
-            {relationValue.map((relatedRecord: any) => (
-              <RecordChip
-                key={relatedRecord.id}
-                objectNameSingular={relationObjectNameSingular}
-                record={relatedRecord}
-              />
-            ))}
-          </StyledRelationChipsContainer>
-        </StyledContainer>
-      </SidePanelProvider>
-    );
-  }
+  const relatedRecords: ObjectRecord[] = isOneToMany
+    ? relationValue
+    : [relationValue];
 
   return (
-    <SidePanelProvider value={{ isInSidePanel }}>
-      <StyledContainer>
-        <StyledRelationChipsContainer>
-          <RecordChip
-            objectNameSingular={relationObjectNameSingular}
-            record={relationValue}
-          />
-        </StyledRelationChipsContainer>
-      </StyledContainer>
-    </SidePanelProvider>
+    <FieldWidgetRelationRecordChips
+      relationRecords={relatedRecords.map((record) => ({
+        record,
+        objectNameSingular: relationObjectNameSingular,
+      }))}
+      isInSidePanel={isInSidePanel}
+    />
   );
 };
