+++ packages/twenty-front/src/modules/page-layout/widgets/field/types/FieldWidgetRelationCardRecord.ts
@@ -1,0 +1,5 @@
+import { type FieldWidgetRelationRecord } from '@/page-layout/widgets/field/types/FieldWidgetRelationRecord';
+
+export type FieldWidgetRelationCardRecord = FieldWidgetRelationRecord & {
+  fieldMetadataId: string;
+};
