+++ packages/twenty-front/src/modules/page-layout/widgets/field/types/FieldWidgetRelationRecord.ts
@@ -1,0 +1,6 @@
+import { type ObjectRecord } from '@/object-record/types/ObjectRecord';
+
+export type FieldWidgetRelationRecord = {
+  record: ObjectRecord;
+  objectNameSingular: string;
+};
